split airbase tests
